# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BELKA AI Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /agents:
    get:
      operationId: listAgents
      tags: [agents]
      summary: List all agents
      responses:
        "200":
          description: List of agents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentList"
    post:
      operationId: createAgent
      tags: [agents]
      summary: Create a new agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgentRequest"
      responses:
        "201":
          description: Agent created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
  /agents/{id}:
    get:
      operationId: getAgent
      tags: [agents]
      summary: Get an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    put:
      operationId: updateAgent
      tags: [agents]
      summary: Update an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgentRequest"
      responses:
        "200":
          description: Updated agent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Agent"
    delete:
      operationId: deleteAgent
      tags: [agents]
      summary: Delete an agent
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
  /admin/models:
    get:
      operationId: listModels
      tags: [admin]
      summary: List AI models
      responses:
        "200":
          description: Models
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModelList"
    post:
      operationId: addModel
      tags: [admin]
      summary: Add an AI model
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddModelRequest"
      responses:
        "201":
          description: Model added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiModel"
  /admin/models/{id}:
    put:
      operationId: updateModel
      tags: [admin]
      summary: Update a model
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddModelRequest"
      responses:
        "200":
          description: Updated model
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiModel"
    delete:
      operationId: deleteModel
      tags: [admin]
      summary: Delete a model
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        avatar:
          type: string
        role:
          type: string
          enum: [coder, reviewer, researcher, designer, orchestrator]
        modelId:
          type: string
        systemPrompt:
          type: string
        capabilities:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        memoryEnabled:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - description
        - role
        - isActive
        - memoryEnabled
        - createdAt
    AgentList:
      type: object
      properties:
        agents:
          type: array
          items:
            $ref: "#/components/schemas/Agent"
      required:
        - agents
    CreateAgentRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        avatar:
          type: string
        role:
          type: string
          enum: [coder, reviewer, researcher, designer, orchestrator]
        modelId:
          type: string
        systemPrompt:
          type: string
        capabilities:
          type: array
          items:
            type: string
        memoryEnabled:
          type: boolean
      required:
        - name
        - description
        - role
    UpdateAgentRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        avatar:
          type: string
        modelId:
          type: string
        systemPrompt:
          type: string
        capabilities:
          type: array
          items:
            type: string
        isActive:
          type: boolean
        memoryEnabled:
          type: boolean
...
    AiModel:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        provider:
          type: string
          enum: [anthropic, openai, google, custom]
        modelId:
          type: string
        apiKey:
          type: string
        isActive:
          type: boolean
        capabilities:
          type: array
          items:
            type: string
        contextWindow:
          type: integer
        costPerToken:
          type: number
        createdAt:
          type: string
      required:
        - id
        - name
        - provider
        - modelId
        - isActive
        - createdAt
    ModelList:
      type: object
      properties:
        models:
          type: array
          items:
            $ref: "#/components/schemas/AiModel"
      required:
        - models
    AddModelRequest:
      type: object
      properties:
        name:
          type: string
        provider:
          type: string
          enum: [anthropic, openai, google, custom]
        modelId:
          type: string
        apiKey:
          type: string
        capabilities:
          type: array
          items:
            type: string
        contextWindow:
          type: integer
        costPerToken:
          type: number
      required:
        - name
        - provider
        - modelId
...